+++ src/config.rs
@@ -23,7 +23,8 @@ use serde;
 use static_analysis::manifest::Permission;
 
 use error::*;
-use {Criticality, print_warning};
+use criticality::Criticality;
+use print_warning;
 
 /// Largest number of threads allowed.
 const MAX_THREADS: i64 = u8::MAX as i64;
